Move utilities file
