+++ Tribler/Test/test_torrent.py
@@ -5,7 +5,7 @@ from twisted.internet.defer import inlineCallbacks
 
 import libtorrent
 
-from Tribler.Core.download.utils import create_torrent_file
+from Tribler.Core.download.utilities import create_torrent_file
 from Tribler.Test.common import TESTS_DATA_DIR
 from Tribler.Test.test_as_server import AbstractServer
 from Tribler.dispersy.util import blocking_call_on_reactor_thread
